+++ kernel/include/kernel/drivers/net/ethernet.h
@@ -3,7 +3,7 @@
 
 #include <kernel/drivers/net/net_types.h>
 
-int eth_send(netif_t *netif, mac_addr_t dest, uint16_t ethertype, uint8_t *data, uint32_t len);
-void eth_receive(netif_t *netif, uint8_t *frame, uint32_t len);
+int eth_send(netif_t *netif, mac_addr_t dest, uint16_t ethertype, uint8_t *data, uint64_t len);
+void eth_receive(netif_t *netif, uint8_t *frame, uint64_t len);
 
 #endif
