+++ kernel/include/kernel/drivers/net/http.h
@@ -15,33 +15,33 @@
 
 typedef struct {
     int status_code;
-    uint32_t content_length;
+    uint64_t content_length;
     char content_type[64];
     char location[512];
     uint8_t *body;
-    uint32_t body_len;
+    uint64_t body_len;
     uint8_t *raw_headers;
-    uint32_t raw_headers_len;
+    uint64_t raw_headers_len;
 } http_response_t;
 
-int http_get(const char *host, uint16_t port, const char *path, http_response_t *response, uint32_t timeout_ms);
-int http_get_tls(const char *host, uint16_t port, const char *path, http_response_t *response, uint32_t timeout_ms, int use_tls);
-int http_get_ip(ipv4_addr_t ip, uint16_t port, const char *host, const char *path, http_response_t *response, uint32_t timeout_ms);
-int http_get_ip_tls(ipv4_addr_t ip, uint16_t port, const char *host, const char *path, http_response_t *response, uint32_t timeout_ms, int use_tls);
+int http_get(const char *host, uint16_t port, const char *path, http_response_t *response, uint64_t timeout_ms);
+int http_get_tls(const char *host, uint16_t port, const char *path, http_response_t *response, uint64_t timeout_ms, int use_tls);
+int http_get_ip(ipv4_addr_t ip, uint16_t port, const char *host, const char *path, http_response_t *response, uint64_t timeout_ms);
+int http_get_ip_tls(ipv4_addr_t ip, uint16_t port, const char *host, const char *path, http_response_t *response, uint64_t timeout_ms, int use_tls);
 int http_post_ip(ipv4_addr_t ip, uint16_t port, const char *host, const char *path,
-                 const char *content_type, const uint8_t *body, uint32_t body_len,
-                 http_response_t *response, uint32_t timeout_ms);
+                 const char *content_type, const uint8_t *body, uint64_t body_len,
+                 http_response_t *response, uint64_t timeout_ms);
 int http_post(const char *host, uint16_t port, const char *path,
-              const char *content_type, const uint8_t *body, uint32_t body_len,
-              http_response_t *response, uint32_t timeout_ms);
+              const char *content_type, const uint8_t *body, uint64_t body_len,
+              http_response_t *response, uint64_t timeout_ms);
 void http_response_free(http_response_t *response);
-int http_download(const char *url, uint8_t *buffer, uint32_t buffer_size, uint32_t *out_size, int *out_status);
-int http_download_ex(const char *url, uint8_t *buffer, uint32_t buffer_size,
-                     uint32_t *out_size, int *out_status, int max_redirects,
-                     uint8_t *headers_buf, uint32_t headers_buf_size, uint32_t *out_headers_len);
+int http_download(const char *url, uint8_t *buffer, uint64_t buffer_size, uint64_t *out_size, int *out_status);
+int http_download_ex(const char *url, uint8_t *buffer, uint64_t buffer_size,
+                     uint64_t *out_size, int *out_status, int max_redirects,
+                     uint8_t *headers_buf, uint64_t headers_buf_size, uint64_t *out_headers_len);
 int http_post_download(const char *url, const char *content_type,
-                       const uint8_t *post_body, uint32_t post_body_len,
-                       uint8_t *buffer, uint32_t buffer_size,
-                       uint32_t *out_size, int *out_status);
+                       const uint8_t *post_body, uint64_t post_body_len,
+                       uint8_t *buffer, uint64_t buffer_size,
+                       uint64_t *out_size, int *out_status);
 
 #endif
