+++ kernel/include/kernel/drivers/net/icmpv6.h
@@ -3,8 +3,8 @@
 
 #include <kernel/drivers/net/net_types.h>
 
-void icmpv6_receive(netif_t *netif, ipv6_addr_t *src, uint8_t *data, uint32_t len);
-int icmpv6_send_echo_request(netif_t *netif, ipv6_addr_t dest, uint16_t id, uint16_t seq, uint8_t *data, uint32_t len);
+void icmpv6_receive(netif_t *netif, ipv6_addr_t *src, uint8_t *data, uint64_t len);
+int icmpv6_send_echo_request(netif_t *netif, ipv6_addr_t dest, uint16_t id, uint16_t seq, uint8_t *data, uint64_t len);
 int icmpv6_send_neighbor_solicitation(netif_t *netif, ipv6_addr_t target);
 int icmpv6_send_router_solicitation(netif_t *netif);
