+++ kernel/include/kernel/rtc.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void rtc_init(void);
-uint32_t rtc_get_time(void);
+uint64_t rtc_get_time(void);
 
 #endif
